+++ src/php/libs/auth.php
@@ -4,46 +4,68 @@ namespace lib;
 
 use db\UserQuery;
 use model\UserModel;
+use Throwable;
 
 class Auth{
 
     public static function login($id, $pwd){
-        $is_success = false;
+
+        try {
+
+            $is_success = false;
     
-        $user = UserQuery::fetchById($id);
-    
-        if(!empty($user) && $user->del_flg !== 1){
-    
-            if(password_verify($pwd, $user->pwd)){
-                $is_success = true;
-                UserModel::setSession($user);
+            $user = UserQuery::fetchById($id);
+        
+            if(!empty($user) && $user->del_flg !== 1){
+        
+                if(password_verify($pwd, $user->pwd)){
+                    $is_success = true;
+                    UserModel::setSession($user);
+                }else{
+                    echo 'パスワードが一致しません。';
+                }
             }else{
-                echo 'パスワードが一致しません。';
+                echo 'ユーザが見つかりません。';
             }
-        }else{
-            echo 'ユーザが見つかりません。';
+            //var_dump($result);
+
+        } catch (Throwable $e) {
+
+            $is_success = false;
+            Msg::push(Msg::DEBUG, $e->getMessage());
+            Msg::push(Msg::ERROR, 'ログイン処理でエラーが発生しました。少し時間をおいてから再度お試しください。');
         }
-        //var_dump($result);
-    
+
         return $is_success;
+
     }
 
     public static function regist($user){
-        $is_success = false;
+
+        try {
+
+            $is_success = false;
     
-        $exist_user = UserQuery::fetchById($user->id);
+            $exist_user = UserQuery::fetchById($user->id);
 
-        if(!empty($exist_user)){
-            echo 'ユーザがすでに存在します。';
-            return false;
-        }
+            if(!empty($exist_user)){
+                echo 'ユーザがすでに存在します。';
+                return false;
+            }
 
-        $is_success = UserQuery::insert($user);
+            $is_success = UserQuery::insert($user);
 
-        if ($is_success) {
+            if ($is_success) {
 
-            UserModel::setSession($user);
+                UserModel::setSession($user);
 
+            }
+
+        } catch (Throwable $e) {
+
+            $is_success = false;
+            Msg::push(Msg::DEBUG, $e->getMessage());
+            Msg::push(Msg::ERROR, 'ユーザー登録でエラーが発生しました。少し時間をおいてから再度お試しください。');
         }
 
         return $is_success;
@@ -51,7 +73,16 @@ class Auth{
 
     public static function isLogin(){
 
-        $user = UserModel::getSession();
+        try {
+
+            $user = UserModel::getSession();
+            
+        } catch (Throwable $e) {
+
+            UserModel::clearSession();
+            Msg::push(Msg::DEBUG, $e->getMessage());
+            return false;
+        }
         
         if(isset($user)) {
             return true;
