+++ src/php/libs/message.php
@@ -3,6 +3,7 @@
 namespace lib;
 
 use model\AbstractModel;
+use Throwable;
 
 class Msg extends AbstractModel
 {
@@ -26,17 +27,27 @@ class Msg extends AbstractModel
     public static function flush()
     {
 
-        $msgs_with_type = static::getSessionAndFlush() ?? [];
+        try {
 
-        foreach ($msgs_with_type as $type => $msgs) {
-            if($type === static::DEBUG && !DEBUG) {
-                continue;
+            $msgs_with_type = static::getSessionAndFlush() ?? [];
+
+            foreach($msgs_with_type as $type => $msgs) {
+                if($type === static::DEBUG && !DEBUG) {
+                    continue;
+                }
+    
+                foreach($msgs as $msg) {
+                    echo "<div>{$type}:{$msg}</div>";
+                }
             }
 
-            foreach ($msgs as $msg) {
-                echo "<div>{$type}:{$msg}</div>";
-            }
+        } catch (Throwable $e) {
+
+            Msg::push(Msg::DEBUG, $e->getMessage());
+            Msg::push(Msg::DEBUG, 'Msg::Flushで例外が発生しました。');
+            
         }
+
     }
